+++ CubicInterpolation/BicubicSplines.h
@@ -29,7 +29,7 @@ public:
    * @brief Properties of an *2-dim* interpolation object.
    */
   struct Definition {
-    std::function<float(float, float)> f;      // function to evaluate
+    std::function<double(double, double)> f;      // function to evaluate
     std::unique_ptr<Axis> f_trafo = nullptr;   // trafo of function values
     std::array<std::unique_ptr<Axis>, N> axis; // trafo of axis
 
@@ -50,19 +50,19 @@ public:
    * value with no knowledge about the transformation which might has choosen.
    * Requires an iterable container with the *x_i* values stored.
    */
-  float evaluate(float x0, float x1) const;
+  double evaluate(double x0, double x1) const;
 
   template <typename T> auto evaluate(T iterable) const {
     return evaluate(iterable[0], iterable[1]);
   }
 
-  std::array<float, 2> prime(float x0, float x1) const;
+  std::array<double, 2> prime(double x0, double x1) const;
 
   template <typename T> auto prime(T iterable) const {
     return prime(iterable[0], iterable[1]);
   }
 
-  float double_prime(float x0, float x1) const;
+  double double_prime(double x0, double x1) const;
 
   template <typename T> auto double_prime(T iterable) const {
     return double_prime(iterable[0], iterable[1]);
